refactor(neuron-ui): refactor settings
